SkillAreas updates classes Service Controller, e DTO

+++ src/main/java/senai/tcc/zupiapi/zupibackend/repositories/SkillAreaRepository.java
@@ -3,6 +3,6 @@ package senai.tcc.zupiapi.zupibackend.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import senai.tcc.zupiapi.zupibackend.model.SkillArea;
 
-public interface SkillAreaRepository extends JpaRepository<SkillArea, Integer> {
+public interface SkillAreaRepository extends JpaRepository<SkillArea, Long> {
     SkillArea findByName(String name);
 }
